Sort existing attachments

+++ resources/assets/src/Editor/Attachments/Attachment.js
@@ -6,24 +6,32 @@ import css from './style.css'
 type Props = {
 	uuid: string,
 	title: string,
-	onDelete: (uuid: string) => any,
+	onDelete?: (uuid: string) => any,
 }
+
+export const DeleteButton = ({
+	onDelete,
+}: {
+	onDelete: (uuid: string) => void,
+}) => (
+	<div onClick={onDelete} className={css.delete}>
+		<i className="fas fa-trash"></i>
+	</div>
+)
 
 class Attachment extends React.Component<Props> {
 	handleDelete = () => {
-		this.props.onDelete(this.props.uuid)
+	    if (this.props.onDelete) this.props.onDelete(this.props.uuid)
 	}
 
 	render() {
 		const { title } = this.props
 
 		return (
-			<li className={css.attachment}>
-				<div onClick={this.handleDelete} className={css.delete}>
-					<i className="fas fa-trash"></i>
-				</div>{' '}
+			<div className={css.attachment}>
+				{this.props.onDelete && <DeleteButton onDelete={this.handleDelete} />}{' '}
 				<div>{title}</div>
-			</li>
+			</div>
 		)
 	}
 }
